+++ packages/sdk/src/AccelbyteSDK.ts
@@ -16,7 +16,7 @@ import { DataRetrievalApi } from '@accelbyte/sdk/api/gdpr/DataRetrievalApi'
 import { InputValidationsApi } from '@accelbyte/sdk/api/iam/InputValidationsApi'
 import { OAuthApi } from '@accelbyte/sdk/api/iam/OAuthApi'
 import { ThirdPartyCredentialApi } from '@accelbyte/sdk/api/iam/ThirdPartyCredentialApi'
-import { TwoFA } from '@accelbyte/sdk/api/iam/TwoFA'
+import { TwoFAApi } from '@accelbyte/sdk/api/iam/TwoFAApi'
 import { UserApi } from '@accelbyte/sdk/api/iam/UserApi'
 import { UserAuthorizationApi } from '@accelbyte/sdk/api/iam/UserAuthorizationApi'
 import { AgreementApi } from '@accelbyte/sdk/api/legal/AgreementApi'
@@ -32,33 +32,30 @@ import { OrderApi } from '@accelbyte/sdk/api/platform/OrderApi'
 import { PaymentApi } from '@accelbyte/sdk/api/platform/PaymentApi'
 import { SubscriptionApi } from '@accelbyte/sdk/api/platform/SubscriptionApi'
 import { WalletApi } from '@accelbyte/sdk/api/platform/WalletApi'
+import { LogLevel } from '@accelbyte/sdk/constants/BuildInfoApp'
 import { Method } from 'axios'
 
 export type Overrides = { config?: SDKRequestConfig; cache?: boolean }
-export type ServiceVersion = { title: string; name: string; version: string | undefined; buildDate: string }
 
 export interface AccelbyteSDK {
   IAM: {
     UserAuthorization(overrides?: Overrides): UserAuthorizationApi
     User(overrides?: Overrides): UserApi
     OAuth(overrides?: Overrides): OAuthApi
-    InputValidation(overrides?: Overrides): InputValidationsApi
+    InputValidations(overrides?: Overrides): InputValidationsApi
     ThirdPartyCredential(overrides?: Overrides): ThirdPartyCredentialApi
-    TwoFA(overrides?: Overrides): TwoFA
-    version: ServiceVersion
+    TwoFA(overrides?: Overrides): TwoFAApi
   }
   BuildInfo: {
     Downloader(overrides?: Overrides): DownloaderApi
     DLC(overrides?: Overrides): DlcApi
     Caching(overrides?: Overrides): CachingApi
-    version: ServiceVersion
   }
   Basic: {
     Misc(overrides?: Overrides): MiscApi
     UserProfile(overrides?: Overrides): UserProfileApi
     FileUpload(overrides?: Overrides): FileUploadApi
     Namespace(overrides?: Overrides): NamespaceApi
-    version: ServiceVersion
   }
   Platform: {
     Currency(overrides?: Overrides): CurrencyApi
@@ -69,29 +66,23 @@ export interface AccelbyteSDK {
     Payment(overrides?: Overrides): PaymentApi
     Subscription(overrides?: Overrides): SubscriptionApi
     Wallet(overrides?: Overrides): WalletApi
-    version: ServiceVersion
   }
   Legal: {
     Eligibilities(overrides?: Overrides): EligibilitiesApi
     Policies(overrides?: Overrides): PoliciesApi
     Agreement(overrides?: Overrides): AgreementApi
     LocalizedPolicyVersions(overrides?: Overrides): LocalizedPolicyVersionsApi
-    version: ServiceVersion
   }
   GDPR: {
     DataDeletion(overrides?: Overrides): DataDeletionApi
     DataRetrieval(overrides?: Overrides): DataRetrievalApi
-    version: ServiceVersion
   }
   Event: {
     Event(overrides?: Overrides): EventApi
-    version: ServiceVersion
   }
   AccelbyteConfig: {
     PublicTemplate<ConfigKeysEnum extends string>(overrides?: Overrides): PublicTemplateApi<ConfigKeysEnum>
-    version: ServiceVersion
   }
-  version: () => void
 
   refreshTokens(newAccessToken: string | undefined | null, newRefreshToken?: string | undefined | null)
 }
@@ -104,6 +95,7 @@ export interface SDKOptions {
 
   // Optional args
   cache?: boolean
+  loglevel?: keyof typeof LogLevel
 }
 
 export interface SDKEvents {