+++ packages/sdk/src/AccelbyteSdkFactory.ts
@@ -4,19 +4,12 @@
  * and restrictions contact your company contract manager.
  */
 import { ApiFactory } from '@accelbyte/sdk/ApiFactory'
+import { LogLevel } from '@accelbyte/sdk/constants/BuildInfoApp'
 import { injectAuthInterceptors } from '@accelbyte/sdk/interceptors/AuthInterceptors'
 import { injectErrorInterceptors } from '@accelbyte/sdk/interceptors/ErrorInterceptor'
 import { AccelbyteSDK, Overrides, SDKEvents, SDKOptions, SDKRequestConfig } from './AccelbyteSDK'
+import buildInfo from './buildInfo.json'
 import { Logger } from './utils/Logger'
-import BasicVersion from '@accelbyte/sdk/generated-public/basic/Version'
-import BuildinfoVersion from '@accelbyte/sdk/generated-public/buildinfo/Version'
-import EventVersion from '@accelbyte/sdk/generated-public/event/Version'
-import GdprVersion from '@accelbyte/sdk/generated-public/gdpr/Version'
-import IamVersion from '@accelbyte/sdk/generated-public/iam/Version'
-import LegalVersion from '@accelbyte/sdk/generated-public/legal/Version'
-import OdinConfigVersion from '@accelbyte/sdk/generated-public/odin-config/Version'
-import PlatformVersion from '@accelbyte/sdk/generated-public/platform/Version'
-import { Network } from '@accelbyte/sdk/utils/Network'
 
 /**
  * This is the main SDK class
@@ -40,6 +33,7 @@ class AccelbyteSDKFactory {
 
   constructor(options: SDKOptions, config?: SDKRequestConfig, events?: SDKEvents) {
     this.options = {
+      loglevel: 'INFO',
       cache: false,
       ...options
     }
@@ -54,7 +48,9 @@ class AccelbyteSDKFactory {
         ...config?.headers
       }
     }
-    // Logger.info('Accelbyte-SDK initialized with config', this.config)
+    if (options.loglevel === LogLevel.DEBUG) {
+      Logger.info('Accelbyte-SDK initialized with config', this.config)
+    }
   }
 
   init(): AccelbyteSDK {
@@ -70,25 +66,22 @@ class AccelbyteSDKFactory {
           ApiFactory.userAuthorization(this.config, this.options, this.refreshToken, this.options.namespace, this.override(overrides)),
         User: (overrides?: Overrides) => ApiFactory.userApi(this.config, this.options.namespace, this.override(overrides)),
         OAuth: (overrides?: Overrides) => ApiFactory.oauthApi(this.config, this.options, this.options.namespace, this.override(overrides)),
-        InputValidation: (overrides?: Overrides) =>
-          ApiFactory.inputValidationApi(this.config, this.options.namespace, this.override(overrides)),
+        InputValidations: (overrides?: Overrides) =>
+          ApiFactory.inputValidationsApi(this.config, this.options.namespace, this.override(overrides)),
         ThirdPartyCredential: (overrides?: Overrides) =>
           ApiFactory.thirdPartyCredentialApi(this.config, this.options.namespace, this.override(overrides)),
-        TwoFA: (overrides?: Overrides) => ApiFactory.twoFA(this.config, this.options.namespace, this.override(overrides)),
-        version: IamVersion
+        TwoFA: (overrides?: Overrides) => ApiFactory.twoFAApi(this.config, this.options.namespace, this.override(overrides))
       },
       BuildInfo: {
         Downloader: (overrides?: Overrides) => ApiFactory.downloaderApi(this.config, this.options.namespace, this.override(overrides)),
         Caching: (overrides?: Overrides) => ApiFactory.cachingApi(this.config, this.options.namespace, this.override(overrides)),
-        DLC: (overrides?: Overrides) => ApiFactory.dlcApi(this.config, this.options.namespace, this.override(overrides)),
-        version: BuildinfoVersion
+        DLC: (overrides?: Overrides) => ApiFactory.dlcApi(this.config, this.options.namespace, this.override(overrides))
       },
       Basic: {
         Misc: (overrides?: Overrides) => ApiFactory.miscApi(this.config, this.options.namespace, this.override(overrides)),
         UserProfile: (overrides?: Overrides) => ApiFactory.userProfileApi(this.config, this.options.namespace, this.override(overrides)),
         FileUpload: (overrides?: Overrides) => ApiFactory.fileUploadApi(this.config, this.options.namespace, this.override(overrides)),
-        Namespace: (overrides?: Overrides) => ApiFactory.namespaceApi(this.config, this.options.namespace, this.override(overrides)),
-        version: BasicVersion
+        Namespace: (overrides?: Overrides) => ApiFactory.namespaceApi(this.config, this.options.namespace, this.override(overrides))
       },
       Platform: {
         Currency: (overrides?: Overrides) => ApiFactory.currencyApi(this.config, this.options.namespace, this.override(overrides)),
@@ -98,8 +91,7 @@ class AccelbyteSDKFactory {
         Order: (overrides?: Overrides) => ApiFactory.orderApi(this.config, this.options.namespace, this.override(overrides)),
         Payment: (overrides?: Overrides) => ApiFactory.paymentApi(this.config, this.options.namespace, this.override(overrides)),
         Subscription: (overrides?: Overrides) => ApiFactory.subscriptionApi(this.config, this.options.namespace, this.override(overrides)),
-        Wallet: (overrides?: Overrides) => ApiFactory.walletApi(this.config, this.options.namespace, this.override(overrides)),
-        version: PlatformVersion
+        Wallet: (overrides?: Overrides) => ApiFactory.walletApi(this.config, this.options.namespace, this.override(overrides))
       },
       Legal: {
         Eligibilities: (overrides?: Overrides) =>
@@ -107,43 +99,18 @@ class AccelbyteSDKFactory {
         Agreement: (overrides?: Overrides) => ApiFactory.agreementApi(this.config, this.options.namespace, this.override(overrides)),
         Policies: (overrides?: Overrides) => ApiFactory.policiesApi(this.config, this.options.namespace, this.override(overrides)),
         LocalizedPolicyVersions: (overrides?: Overrides) =>
-          ApiFactory.localizedPolicyVersionsApi(this.config, this.options.namespace, this.override(overrides)),
-        version: LegalVersion
+          ApiFactory.localizedPolicyVersionsApi(this.config, this.options.namespace, this.override(overrides))
       },
       GDPR: {
         DataDeletion: (overrides?: Overrides) => ApiFactory.dataDeletionApi(this.config, this.options.namespace, this.override(overrides)),
-        DataRetrieval: (overrides?: Overrides) =>
-          ApiFactory.dataRetrievalApi(this.config, this.options.namespace, this.override(overrides)),
-        version: GdprVersion
+        DataRetrieval: (overrides?: Overrides) => ApiFactory.dataRetrievalApi(this.config, this.options.namespace, this.override(overrides))
       },
       Event: {
-        Event: (overrides?: Overrides) => ApiFactory.eventApi(this.config, this.options.namespace, this.override(overrides)),
-        version: EventVersion
+        Event: (overrides?: Overrides) => ApiFactory.eventApi(this.config, this.options.namespace, this.override(overrides))
       },
       AccelbyteConfig: {
         PublicTemplate: (overrides?: Overrides) =>
-          ApiFactory.publicTemplateApi(this.config, this.options.namespace, this.override(overrides)),
-        version: OdinConfigVersion
-      },
-      version: () => {
-        console.log('IamVersion: ', IamVersion.version)
-        console.log('BuildinfoVersion: ', BuildinfoVersion.version)
-        console.log('BasicVersion: ', BasicVersion.version)
-        console.log('PlatformVersion: ', PlatformVersion.version)
-        console.log('LegalVersion: ', LegalVersion.version)
-        console.log('GdprVersion: ', GdprVersion.version)
-        console.log('EventVersion: ', EventVersion.version)
-
-        const axiosInstance = Network.create(this.config)
-        axiosInstance
-          .get('https://development.accelbyte.io/iam/version')
-          .then(res => {
-            console.log('-- axiosInstance res', res.data)
-            console.log('-- current ver', IamVersion)
-          })
-          .catch(err => {
-            console.log('-- axiosInstance err', err)
-          })
+          ApiFactory.publicTemplateApi(this.config, this.options.namespace, this.override(overrides))
       }
     }
   }
@@ -180,8 +147,13 @@ class AccelbyteSDKFactory {
 }
 
 const sdkInit = ({ options, config, onEvents }: { options: SDKOptions; config?: SDKRequestConfig; onEvents?: SDKEvents }): AccelbyteSDK => {
-  Logger.info('Accelbyte-SDK instantiated:', options)
+  if (options.loglevel === LogLevel.DEBUG) {
+    Logger.info('Accelbyte-SDK instantiated:', options)
+  }
   const sdkFactory = new AccelbyteSDKFactory(options, config, onEvents)
+  if (options.loglevel === LogLevel.DEBUG) {
+    Logger.info(`Accelbyte-SDK version: ${buildInfo.version} \nBuild: ${buildInfo.build} \nTimestamp: ${new Date(buildInfo.timestamp)}`)
+  }
   return sdkFactory.init()
 }